fix: now percentile95 use correct formula, percentile and maxResponseSize are doubles
- fixed tests for this logic

+++ src/main/java/academy/stats/ReportTotalStats.java
@@ -6,8 +6,8 @@ import java.util.Set;
 public record ReportTotalStats(
         long totalRequests,
         double averageResponseSize,
-        int maxResponseSize,
-        int percentile95,
+        double maxResponseSize,
+        double percentile95,
         Map<Integer, Long> responseCodes,
         Map<String, Long> mostPopular10Resources,
         Map<String, RequestDateInfo> requestsByDateInPercents,
